requirements.txt
flask==2.0.1
ultralytics==8.0.0
numpy==1.23.5
opencv-python==4.7.0.72
pillow==9.3.0
matplotlib==3.7.1
scikit-image==0.20.0
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0 